Views sub-folders restructure
