+++ src/views/components/NavBar.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import './navbar.css';
+import '../styles.css';
 
 const Navbar = () => {
   return (
