+++ src/views/pages/ExamplesPage.js
