+++ src/views/styles.css
@@ -1,3 +1,5 @@
+/* SIDEBAR STYLE STARTS */
+
 .sidebar {
   position: fixed;
   left: 0;
@@ -31,3 +33,39 @@
   cursor: pointer;
   color: gray;
 }
+/* SIDEBAR STYLES ENDS */
+
+/* NAVBAR STYLES STARTS */
+.navbar {
+  width: 100%;
+  border-bottom: 1px solid #ddd;
+}
+
+.navbar-div {
+  display: flex;
+  width: 80%;
+  justify-content: space-between;
+  align-items: center;
+  margin: 0 auto;
+  background: #fff;
+}
+
+ul {
+  display: inline-flex;
+  font-size: 23px;
+}
+
+a {
+  text-decoration: none;
+}
+
+li {
+  color: #000;
+  list-style-type: none;
+  padding: 15px;
+}
+
+li:hover {
+  color: gray;
+}
+/* NAVBAR STYLESA ENDS */
